Trait linera_storage::Store

pub trait Store: Sized {
    type Context: Context<Extra = ChainRuntimeContext<Self>, Error = Self::ContextError> + Clone + Send + Sync + 'static;
    type ContextError: Error + Debug + Sync + Send;

    fn load_chain<'life0, 'async_trait>(
        &'life0 self,
        id: ChainId
    ) -> Pin<Box<dyn Future<Output = Result<ChainStateView<Self::Context>, ViewError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_value<'life0, 'async_trait>(
        &'life0 self,
        hash: CryptoHash
    ) -> Pin<Box<dyn Future<Output = Result<HashedValue, ViewError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_value<'life0, 'life1, 'async_trait>(
        &'life0 self,
        value: &'life1 HashedValue
    ) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn read_certificate<'life0, 'async_trait>(
        &'life0 self,
        hash: CryptoHash
    ) -> Pin<Box<dyn Future<Output = Result<Certificate, ViewError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_certificate<'life0, 'life1, 'async_trait>(
        &'life0 self,
        certificate: &'life1 Certificate
    ) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn wasm_runtime(&self) -> Option<WasmRuntime>;

    fn load_active_chain<'life0, 'async_trait>(
        &'life0 self,
        id: ChainId
    ) -> Pin<Box<dyn Future<Output = Result<ChainStateView<Self::Context>, ChainError>> + Send + 'async_trait>>
    where
        ChainRuntimeContext<Self>: ExecutionRuntimeContext,
        ViewError: From<Self::ContextError>,
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn read_certificates<'life0, 'async_trait, I>(
        &'life0 self,
        keys: I
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Certificate>, ViewError>> + Send + 'async_trait>>
    where
        Self: Clone + Send + 'static,
        I: 'async_trait + IntoIterator<Item = CryptoHash> + Send,
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn create_chain<'life0, 'async_trait>(
        &'life0 self,
        committee: Committee,
        admin_id: ChainId,
        description: ChainDescription,
        public_key: PublicKey,
        balance: Balance,
        timestamp: Timestamp
    ) -> Pin<Box<dyn Future<Output = Result<(), ChainError>> + Send + 'async_trait>>
    where
        ChainRuntimeContext<Self>: ExecutionRuntimeContext,
        ViewError: From<Self::ContextError>,
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn load_application<'life0, 'life1, 'async_trait>(
        &'life0 self,
        application_description: &'life1 UserApplicationDescription
    ) -> Pin<Box<dyn Future<Output = Result<UserApplicationCode, ExecutionError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Communicate with a persistent storage using the “views” abstraction.

Required Associated Types

type Context: Context<Extra = ChainRuntimeContext<Self>, Error = Self::ContextError> + Clone + Send + Sync + 'static

The low-level storage implementation in use.

type ContextError: Error + Debug + Sync + Send

Alias to provide simpler trait bounds ViewError: From<Self::ContextError>

Required Methods

fn load_chain<'life0, 'async_trait>(
&'life0 self,
id: ChainId
) -> Pin<Box<dyn Future<Output = Result<ChainStateView<Self::Context>, ViewError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Loads the view of a chain state.

fn read_value<'life0, 'async_trait>(
&'life0 self,
hash: CryptoHash
) -> Pin<Box<dyn Future<Output = Result<HashedValue, ViewError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Reads the value with the given hash.

fn write_value<'life0, 'life1, 'async_trait>(
&'life0 self,
value: &'life1 HashedValue
) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Writes the given value.

fn read_certificate<'life0, 'async_trait>(
&'life0 self,
hash: CryptoHash
) -> Pin<Box<dyn Future<Output = Result<Certificate, ViewError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Reads the certificate with the given hash.

fn write_certificate<'life0, 'life1, 'async_trait>(
&'life0 self,
certificate: &'life1 Certificate
) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Writes the given certificate.

fn wasm_runtime(&self) -> Option<WasmRuntime>

Selects the WebAssembly runtime to use for applications (if any).

Provided Methods

fn load_active_chain<'life0, 'async_trait>(
&'life0 self,
id: ChainId
) -> Pin<Box<dyn Future<Output = Result<ChainStateView<Self::Context>, ChainError>> + Send + 'async_trait>>where
ChainRuntimeContext<Self>: ExecutionRuntimeContext,
ViewError: From<Self::ContextError>,
'life0: 'async_trait,
Self: Sync + 'async_trait,

Loads the view of a chain state and checks that it is active.

fn read_certificates<'life0, 'async_trait, I>(
&'life0 self,
keys: I
) -> Pin<Box<dyn Future<Output = Result<Vec<Certificate>, ViewError>> + Send + 'async_trait>>where
Self: Clone + Send + 'static,
I: 'async_trait + IntoIterator<Item = CryptoHash> + Send,
'life0: 'async_trait,
Self: Sync + 'async_trait,

Reads a number of certificates in parallel.

fn create_chain<'life0, 'async_trait>(
&'life0 self,
committee: Committee,
admin_id: ChainId,
description: ChainDescription,
public_key: PublicKey,
balance: Balance,
timestamp: Timestamp
) -> Pin<Box<dyn Future<Output = Result<(), ChainError>> + Send + 'async_trait>>where
ChainRuntimeContext<Self>: ExecutionRuntimeContext,
ViewError: From<Self::ContextError>,
'life0: 'async_trait,
Self: Sync + 'async_trait,

Initializes a chain in a simple way (used for testing and to create a genesis state).

fn load_application<'life0, 'life1, 'async_trait>(
&'life0 self,
application_description: &'life1 UserApplicationDescription
) -> Pin<Box<dyn Future<Output = Result<UserApplicationCode, ExecutionError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

Creates a [linera-sdk::UserApplication] instance using the bytecode in storage referenced by the application_description.