Crate linera_storage

This module defines the storage abstractions for individual chains and certificates.

Re-exports

pub use crate::db_storage::READ_CERTIFICATE_COUNTER;

pub use crate::db_storage::READ_VALUE_COUNTER;

pub use crate::db_storage::WRITE_CERTIFICATE_COUNTER;

pub use crate::db_storage::WRITE_VALUE_COUNTER;

Structs

ChainRuntimeContext

DbStorage

A DbStorage wrapping with Arc

TestClock

A clock implementation that uses a stored number of microseconds and that can be updated explicitly. All clones share the same time, and setting it in one clone updates all the others.

WallClock

A Clock implementation using the system clock.

Traits

Clock

A clock that can be used to get the current Timestamp.

Storage

Communicate with a persistent storage using the “views” abstraction.

Type Aliases

DynamoDbStorage

MemoryStorage

RocksDbStorage

ScyllaDbStorage

ServiceStorage