Crate linera_storage

This module defines the storage abstractions for individual chains and certificates.

Modules

metrics

Structs

ChainRuntimeContext

An implementation of ExecutionRuntimeContext suitable for the core protocol.

ChainStatesFirstAssignment

An implementation of DualStoreRootKeyAssignment that stores the chain states into the first store.

DbStorage

Main implementation of the Storage trait.

TestClock

A clock implementation that uses a stored number of microseconds and that can be updated explicitly. All clones share the same time, and setting it in one clone updates all the others.

WallClock

A Clock implementation using the system clock.

Enums

ResultReadCertificates

The result of processing the obtained read certificates.

Constants

DEFAULT_NAMESPACE

The default namespace to be used when none is specified

Traits

Clock

A clock that can be used to get the current Timestamp.

Storage

Communicate with a persistent storage using the “views” abstraction.