Crate linera_storage

Structs

ChainRuntimeContext

DynamoDbStoreClient

MemoryStoreClient

RocksdbStoreClient

Constants

READ_CERTIFICATE_COUNTER

The metric counting how often a certificate is read from storage.

READ_VALUE_COUNTER

The metric counting how often a value is read from storage.

WRITE_CERTIFICATE_COUNTER

The metric counting how often a certificate is written to storage.

WRITE_VALUE_COUNTER

The metric counting how often a value is written to storage.

Traits

Store

Communicate with a persistent storage using the “views” abstraction.