Trait linera_storage::Storage

pub trait Storage: Sized {
    type Context: Context<Extra = ChainRuntimeContext<Self>, Error = Self::ContextError> + Clone + Send + Sync + 'static;
    type ContextError: Error + Debug + Sync + Send;

    // Required methods
    fn current_time(&self) -> Timestamp;
    fn load_chain<'life0, 'async_trait>(
        &'life0 self,
        id: ChainId
    ) -> Pin<Box<dyn Future<Output = Result<ChainStateView<Self::Context>, ViewError>> + Send + 'async_trait>>
       where ViewError: From<Self::ContextError>,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn contains_value<'life0, 'async_trait>(
        &'life0 self,
        hash: CryptoHash
    ) -> Pin<Box<dyn Future<Output = Result<bool, ViewError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn read_value<'life0, 'async_trait>(
        &'life0 self,
        hash: CryptoHash
    ) -> Pin<Box<dyn Future<Output = Result<HashedValue, ViewError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn read_values_downward<'life0, 'async_trait>(
        &'life0 self,
        from: CryptoHash,
        limit: u32
    ) -> Pin<Box<dyn Future<Output = Result<Vec<HashedValue>, ViewError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn write_value<'life0, 'life1, 'async_trait>(
        &'life0 self,
        value: &'life1 HashedValue
    ) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn write_values<'life0, 'life1, 'async_trait>(
        &'life0 self,
        values: &'life1 [HashedValue]
    ) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn contains_certificate<'life0, 'async_trait>(
        &'life0 self,
        hash: CryptoHash
    ) -> Pin<Box<dyn Future<Output = Result<bool, ViewError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn read_certificate<'life0, 'async_trait>(
        &'life0 self,
        hash: CryptoHash
    ) -> Pin<Box<dyn Future<Output = Result<Certificate, ViewError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn write_certificate<'life0, 'life1, 'async_trait>(
        &'life0 self,
        certificate: &'life1 Certificate
    ) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn write_certificates<'life0, 'life1, 'async_trait>(
        &'life0 self,
        certificate: &'life1 [Certificate]
    ) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn wasm_runtime(&self) -> Option<WasmRuntime>;

    // Provided methods
    fn load_active_chain<'life0, 'async_trait>(
        &'life0 self,
        id: ChainId
    ) -> Pin<Box<dyn Future<Output = Result<ChainStateView<Self::Context>, ChainError>> + Send + 'async_trait>>
       where ChainRuntimeContext<Self>: ExecutionRuntimeContext,
             ViewError: From<Self::ContextError>,
             Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn read_certificates<'life0, 'async_trait, I>(
        &'life0 self,
        keys: I
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Certificate>, ViewError>> + Send + 'async_trait>>
       where Self: Clone + Send + 'static + Sync + 'async_trait,
             I: 'async_trait + IntoIterator<Item = CryptoHash> + Send,
             'life0: 'async_trait { ... }
    fn create_chain<'life0, 'async_trait>(
        &'life0 self,
        committee: Committee,
        admin_id: ChainId,
        description: ChainDescription,
        public_key: PublicKey,
        balance: Amount,
        timestamp: Timestamp
    ) -> Pin<Box<dyn Future<Output = Result<(), ChainError>> + Send + 'async_trait>>
       where ChainRuntimeContext<Self>: ExecutionRuntimeContext,
             ViewError: From<Self::ContextError>,
             Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn load_contract<'life0, 'life1, 'async_trait>(
        &'life0 self,
        application_description: &'life1 UserApplicationDescription
    ) -> Pin<Box<dyn Future<Output = Result<UserContractCode, ExecutionError>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn load_service<'life0, 'life1, 'async_trait>(
        &'life0 self,
        application_description: &'life1 UserApplicationDescription
    ) -> Pin<Box<dyn Future<Output = Result<UserServiceCode, ExecutionError>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }

}

Communicate with a persistent storage using the “views” abstraction.

Required Associated Types

type Context: Context<Extra = ChainRuntimeContext<Self>, Error = Self::ContextError> + Clone + Send + Sync + 'static

The low-level storage implementation in use.

type ContextError: Error + Debug + Sync + Send

Alias to provide simpler trait bounds ViewError: From<Self::ContextError>

Required Methods

fn current_time(&self) -> Timestamp

Returns the current wall clock time.

fn load_chain<'life0, 'async_trait>( &'life0 self, id: ChainId ) -> Pin<Box<dyn Future<Output = Result<ChainStateView<Self::Context>, ViewError>> + Send + 'async_trait>>where ViewError: From<Self::ContextError>, Self: 'async_trait, 'life0: 'async_trait,

Loads the view of a chain state.

fn contains_value<'life0, 'async_trait>( &'life0 self, hash: CryptoHash ) -> Pin<Box<dyn Future<Output = Result<bool, ViewError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Tests existence of a value with the given hash.

fn read_value<'life0, 'async_trait>( &'life0 self, hash: CryptoHash ) -> Pin<Box<dyn Future<Output = Result<HashedValue, ViewError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Reads the value with the given hash.

fn read_values_downward<'life0, 'async_trait>( &'life0 self, from: CryptoHash, limit: u32 ) -> Pin<Box<dyn Future<Output = Result<Vec<HashedValue>, ViewError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Reads the values in descending order from the given hash.

fn write_value<'life0, 'life1, 'async_trait>( &'life0 self, value: &'life1 HashedValue ) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Writes the given value.

fn write_values<'life0, 'life1, 'async_trait>( &'life0 self, values: &'life1 [HashedValue] ) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Writes several values

fn contains_certificate<'life0, 'async_trait>( &'life0 self, hash: CryptoHash ) -> Pin<Box<dyn Future<Output = Result<bool, ViewError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Tests existence of the certificate with the given hash.

fn read_certificate<'life0, 'async_trait>( &'life0 self, hash: CryptoHash ) -> Pin<Box<dyn Future<Output = Result<Certificate, ViewError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Reads the certificate with the given hash.

fn write_certificate<'life0, 'life1, 'async_trait>( &'life0 self, certificate: &'life1 Certificate ) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Writes the given certificate.

fn write_certificates<'life0, 'life1, 'async_trait>( &'life0 self, certificate: &'life1 [Certificate] ) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Writes a vector of certificates.

fn wasm_runtime(&self) -> Option<WasmRuntime>

Selects the WebAssembly runtime to use for applications (if any).

Provided Methods

fn load_active_chain<'life0, 'async_trait>( &'life0 self, id: ChainId ) -> Pin<Box<dyn Future<Output = Result<ChainStateView<Self::Context>, ChainError>> + Send + 'async_trait>>where ChainRuntimeContext<Self>: ExecutionRuntimeContext, ViewError: From<Self::ContextError>, Self: Sync + 'async_trait, 'life0: 'async_trait,

Loads the view of a chain state and checks that it is active.

fn read_certificates<'life0, 'async_trait, I>( &'life0 self, keys: I ) -> Pin<Box<dyn Future<Output = Result<Vec<Certificate>, ViewError>> + Send + 'async_trait>>where Self: Clone + Send + 'static + Sync + 'async_trait, I: 'async_trait + IntoIterator<Item = CryptoHash> + Send, 'life0: 'async_trait,

Reads a number of certificates in parallel.

fn create_chain<'life0, 'async_trait>( &'life0 self, committee: Committee, admin_id: ChainId, description: ChainDescription, public_key: PublicKey, balance: Amount, timestamp: Timestamp ) -> Pin<Box<dyn Future<Output = Result<(), ChainError>> + Send + 'async_trait>>where ChainRuntimeContext<Self>: ExecutionRuntimeContext, ViewError: From<Self::ContextError>, Self: Sync + 'async_trait, 'life0: 'async_trait,

Initializes a chain in a simple way (used for testing and to create a genesis state).

fn load_contract<'life0, 'life1, 'async_trait>( &'life0 self, application_description: &'life1 UserApplicationDescription ) -> Pin<Box<dyn Future<Output = Result<UserContractCode, ExecutionError>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Creates a UserContractCode instance using the bytecode in storage referenced by the application_description.

fn load_service<'life0, 'life1, 'async_trait>( &'life0 self, application_description: &'life1 UserApplicationDescription ) -> Pin<Box<dyn Future<Output = Result<UserServiceCode, ExecutionError>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Creates a [linera-sdk::UserContract] instance using the bytecode in storage referenced by the application_description.

Object Safety

This trait is not object safe.